Move it to GPT2 tokenization tests.
